fix double routes folder?
